visual-python/requirements/dev.txt
# Development & local build requirements
-r user.txt
-r build.txt
robotframework==7.0.0
robotframework-seleniumlibrary==6.2.0
pytest==8.1.1
